fixing parsing issue
