+++ cortex/_lib/__init__.py
@@ -7,7 +7,7 @@ import logging
 import pprint
 
 from . import config, exp, log_utils, models
-from .parsing import default_args, parse_args
+from .parsing import default_args, parse_args, update_args
 from .viz import init as viz_init
 
 __author__ = 'R Devon Hjelm'
@@ -49,7 +49,7 @@ def setup_experiment(args, model=None):
         model_name = model.__class__.__name__
 
     experiment_args = copy.deepcopy(default_args)
-    exp.update_args(experiment_args)
+    update_args(experiment_args, exp.ARGS)
 
     viz_init(config.CONFIG.viz)
 
@@ -58,8 +58,8 @@ def setup_experiment(args, model=None):
                    args.out_path, args.clean, config.CONFIG)
     else:
         name = args.name or model_name
-        exp.setup_new(model.defaults, name, args.out_path, args.clean,
-                      config.CONFIG, args.load_models, args.reloads)
+        exp.setup_new(name, args.out_path, args.clean, config.CONFIG,
+                      args.load_models, args.reloads)
 
     exp.configure_from_yaml(config_file=args.config_file)
 
@@ -86,7 +86,8 @@ def setup_experiment(args, model=None):
 
     update_nested_dicts(command_line_args['model'], model.kwargs)
     command_line_args['model'].update(**model.kwargs)
-    exp.update_args(command_line_args)
+    update_args(command_line_args, exp.ARGS)
+
     for k, v in exp.ARGS.items():
         logger.info('Ultimate {} arguments: \n{}'
                     .format(k, pprint.pformat(v)))
